requirements.txt
# ============================================================
# Vigour POC — Python 3.11
# GPU machine: install onnxruntime-gpu and paddlepaddle-gpu
# CPU machine: install onnxruntime and paddlepaddle
# ============================================================

# ---- Computer Vision core ----
opencv-python-headless>=4.9.0
numpy>=1.26.0

# ---- Person detection ----
ultralytics>=8.1.0           # YOLOv8 (pt + ONNX export)
onnxruntime-gpu>=1.17.0      # ONNX inference on GPU (swap for onnxruntime on CPU)

# ---- Multi-person tracking ----
bytetracker>=0.1.4

# ---- Pose estimation ----
mmpose>=1.3.0
openmim>=0.3.9
# mmcv and mmengine installed via:  mim install mmcv mmengine

# ---- Bib OCR ----
paddlepaddle-gpu>=2.6.0      # swap for paddlepaddle on CPU
paddleocr>=2.7.3

# ---- API ----
fastapi[standard]>=0.110.0
uvicorn[standard]>=0.29.0

# ---- Task queue ----
celery>=5.3.6
redis>=5.0.3

# ---- Database ----
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9

# ---- Cloud ----
google-cloud-storage>=2.16.0

# ---- Dev / notebooks ----
jupyter>=1.0.0
matplotlib>=3.8.3
pandas>=2.2.1
tqdm>=4.66.2

# ---- Testing ----
pytest>=8.1.0
pytest-cov>=5.0.0
